CRecursiveMutex: move to platform/posix/threads
