+++ xbmc/platform/posix/threads/CMakeLists.txt
@@ -1,4 +1,6 @@
-set(SOURCES ThreadImplPosix.cpp)
-set(HEADERS ThreadImplPosix.h)
+set(SOURCES RecursiveMutex.cpp
+            ThreadImplPosix.cpp)
+set(HEADERS RecursiveMutex.h
+            ThreadImplPosix.h)
 
 core_add_library(platform_posix_threads)
